Ditch multiboot and write our own MBR bootstrap
I am still compiling everything except the MBR sector into a big ELF
blob, and then carving out the text section of that ELF to make the
final disk image. The reason for this is so that we have debug symbols
after we jump into boot32.asm (formerly boot.asm). Need to find a
less hacky way to do this.