Error handling for read sectors
Improved comments for the read sectors operation, and error checking in
case something goes wrong. In the process discovered that the int 0x13
was always returning a failure (but we were forging on anyway) because
the disk did not have 9 sectors (one for MBR, 8 for the first real
code), so changed Makefile to ensure we pad with zeroes.